Entrega 3

+++ springboot_backendmg/src/main/java/com/backendmg/springboot_backendmg/dto/DetalleOrdenRequest.java
@@ -3,12 +3,12 @@ package com.backendmg.springboot_backendmg.dto;
 public class DetalleOrdenRequest {
     private String nombreProducto;
     private Integer cantidad;
-    private Integer precioUnitario;
+    private Double precioUnitario;
 
     public String getNombreProducto() { return nombreProducto; }
     public void setNombreProducto(String nombreProducto) { this.nombreProducto = nombreProducto; }
     public Integer getCantidad() { return cantidad; }
     public void setCantidad(Integer cantidad) { this.cantidad = cantidad; }
-    public Integer getPrecioUnitario() { return precioUnitario; }
-    public void setPrecioUnitario(Integer precioUnitario) { this.precioUnitario = precioUnitario; }
+    public Double getPrecioUnitario() { return precioUnitario; }
+    public void setPrecioUnitario(Double precioUnitario) { this.precioUnitario = precioUnitario; }
 }
